Align to BlueST-SDK-V1.2.18

+++ st_blue_sdk/src/main/java/com/st/blue_sdk/features/extended/medical_signal/MedicalSignal16BitFeature.kt
@@ -40,7 +40,7 @@ class MedicalSignal16BitFeature(
         if(signalType.precision == MedicalPrecision.UBIT16) {
             for (i in 0 until numberOfSamples) {
                 values.add(
-                    NumberConversion.LittleEndian.bytesToUInt16(data, dataOffset + 5 + i * 2).toInt()
+                    NumberConversion.LittleEndian.bytesToUInt16(data, dataOffset + 5 + i * 2)
                 )
             }
         } else {
