+++ st_blue_sdk/src/main/java/com/st/blue_sdk/features/extended/piano/Piano.kt
@@ -91,7 +91,7 @@ class Piano(
                         packCommandRequest(
                             featureBit,
                             CommandPianoSound.PIANO_COMMAND_TYPE_START,
-                            byteArrayOf(command.key.toByte())
+                            byteArrayOf(command.key)
                         )
                     }
                 }
